# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Returns Services API",
        "description": "API documentation for Todo items",
        "contact": {
            "email": "test@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Returns Services API"
        }
    ],
    "paths": {
...
        "/todos": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "List of all todo items",
                "operationId": "18e676fc7a1838c48c63274e52066ac8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server error. Try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Todo"
                ],
                "summary": "Create new todo item",
                "description": "Colors are in hex format. There is the list of available colors: #ffffff, #ff0000, #00ff00, #0000ff, #ffff00, #00ffff, #ff00ff, #000000, #808080, #ff8000, #8000ff",
                "operationId": "ae1281faa0d4aef5c4f3e277c6a57e7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Buy milk"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Buy milk"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "favorite": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#ff0000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Buy milk"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Buy milk"
                                        },
                                        "completed": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "favorite": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "color": {
                                            "type": "string",
                                            "example": "#ff0000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/todos/{id}": {
            "get": {
                "tags": [
                    "Todo"
                ],
                "summary": "Get todo item",
                "operationId": "1b99eaebc4bafa42301fbf780b77da93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Buy milk"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Buy milk"
                                        },
                                        "completed": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "favorite": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "color": {
                                            "type": "string",
                                            "example": "#ff0000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Todo"
                ],
                "summary": "Update todo item",
                "description": "Colors are in hex format. There is the list of available colors: #ffffff, #ff0000, #00ff00, #0000ff, #ffff00, #00ffff, #ff00ff, #000000, #808080, #ff8000, #8000ff",
                "operationId": "c10cfb143d3c35afe02449b07608baba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Buy milk"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Buy milk"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "favorite": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#ff0000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Todo"
                ],
                "summary": "Delete todo item",
                "operationId": "e6e24dd784980cf6f9fe389df03fdd14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}